requirements.txt
# RealSense
pyrealsense2>=2.50.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# YOLO
ultralytics>=8.0.200
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0

# NVdia GPU
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# GUI
# tkinter

# pip install -r requirements.txt
